Add vic-machine configure container-network support
This change adds support for container network flags in vic-machine
configure. Any existing container networks and their related settings
must be specified exactly along with the new networks. Adding new
container networks is supported (deleting is not). If an attempt to
change existing network(s) is made, the configure operation will fail.
