+++ cmd/vic-machine/create/create.go
@@ -17,7 +17,6 @@ package create
 import (
 	"context"
 	"crypto/tls"
-	"encoding"
 	"fmt"
 	"io/ioutil"
 	"net"
@@ -41,7 +40,6 @@ import (
 	"github.com/vmware/vic/pkg/certificate"
 	"github.com/vmware/vic/pkg/errors"
 	"github.com/vmware/vic/pkg/flags"
-	"github.com/vmware/vic/pkg/ip"
 	"github.com/vmware/vic/pkg/trace"
 )
 
@@ -98,23 +96,21 @@ type Create struct {
 	clientCAs   cli.StringSlice `arg:"tls-ca"`
 	registryCAs cli.StringSlice `arg:"registry-ca"`
 
-	containerNetworks         cli.StringSlice `arg:"container-network"`
-	containerNetworksGateway  cli.StringSlice `arg:"container-network-gateway"`
-	containerNetworksIPRanges cli.StringSlice `arg:"container-network-ip-range"`
-	containerNetworksDNS      cli.StringSlice `arg:"container-network-dns"`
-	volumeStores              cli.StringSlice `arg:"volume-store"`
-	insecureRegistries        cli.StringSlice `arg:"insecure-registry"`
-	whitelistRegistries       cli.StringSlice `arg:"whitelist-registry"`
-	dns                       cli.StringSlice `arg:"dns-server"`
-	clientNetworkName         string
-	clientNetworkGateway      string
-	clientNetworkIP           string
-	publicNetworkName         string
-	publicNetworkGateway      string
-	publicNetworkIP           string
-	managementNetworkName     string
-	managementNetworkGateway  string
-	managementNetworkIP       string
+	containerNetworks common.CNetworks
+
+	volumeStores             cli.StringSlice `arg:"volume-store"`
+	insecureRegistries       cli.StringSlice `arg:"insecure-registry"`
+	whitelistRegistries      cli.StringSlice `arg:"whitelist-registry"`
+	dns                      cli.StringSlice `arg:"dns-server"`
+	clientNetworkName        string
+	clientNetworkGateway     string
+	clientNetworkIP          string
+	publicNetworkName        string
+	publicNetworkGateway     string
+	publicNetworkIP          string
+	managementNetworkName    string
+	managementNetworkGateway string
+	managementNetworkIP      string
 
 	memoryReservLimits string
 	cpuReservLimits    string
@@ -294,33 +290,9 @@ func (c *Create) Flags() []cli.Flag {
 			Usage:  "DNS server for the client, public, and management networks. Defaults to 8.8.8.8 and 8.8.4.4 when VCH uses static IP",
 			Hidden: true,
 		},
+	}
 
-		// container networks - mapped from vSphere
-		cli.StringSliceFlag{
-			Name:  "container-network, cn",
-			Value: &c.containerNetworks,
-			Usage: "vSphere network list that containers can use directly with labels, e.g. vsphere-net:backend. Defaults to DCHP - see advanced help (-x).",
-		},
-		cli.StringSliceFlag{
-			Name:   "container-network-gateway, cng",
-			Value:  &c.containerNetworksGateway,
-			Usage:  "Gateway for the container network's subnet in CONTAINER-NETWORK:SUBNET format, e.g. vsphere-net:172.16.0.1/16",
-			Hidden: true,
-		},
-		cli.StringSliceFlag{
-			Name:   "container-network-ip-range, cnr",
-			Value:  &c.containerNetworksIPRanges,
-			Usage:  "IP range for the container network in CONTAINER-NETWORK:IP-RANGE format, e.g. vsphere-net:172.16.0.0/24, vsphere-net:172.16.0.10-172.16.0.20",
-			Hidden: true,
-		},
-		cli.StringSliceFlag{
-			Name:   "container-network-dns, cnd",
-			Value:  &c.containerNetworksDNS,
-			Usage:  "DNS servers for the container network in CONTAINER-NETWORK:DNS format, e.g. vsphere-net:8.8.8.8. Ignored if no static IP assigned.",
-			Hidden: true,
-		},
-
-		// memory
+	memory := []cli.Flag{
 		cli.IntFlag{
 			Name:        "memory, mem",
 			Value:       0,
@@ -347,8 +319,9 @@ func (c *Create) Flags() []cli.Flag {
 			Hidden:      true,
 			Destination: &c.MemoryMB,
 		},
+	}
 
-		// cpu
+	cpu := []cli.Flag{
 		cli.IntFlag{
 			Name:        "cpu",
 			Value:       0,
@@ -375,8 +348,9 @@ func (c *Create) Flags() []cli.Flag {
 			Hidden:      true,
 			Destination: &c.NumCPUs,
 		},
+	}
 
-		// TLS
+	tls := []cli.Flag{
 		cli.StringFlag{
 			Name:        "tls-cname",
 			Value:       "",
@@ -434,8 +408,9 @@ func (c *Create) Flags() []cli.Flag {
 			Destination: &c.keySize,
 			Hidden:      true,
 		},
+	}
 
-		// registries
+	registries := []cli.Flag{
 		cli.StringSliceFlag{
 			Name:   "registry-ca, rc",
 			Usage:  "Specify a list of additional certificate authority files to use to verify secure registry servers",
@@ -506,10 +481,11 @@ func (c *Create) Flags() []cli.Flag {
 	iso := c.ImageFlags(true)
 	debug := c.DebugFlags()
 	proxies := c.proxies.ProxyFlags(true)
+	cNetwork := c.containerNetworks.CNetworkFlags(true)
 
 	// flag arrays are declared, now combined
 	var flags []cli.Flag
-	for _, f := range [][]cli.Flag{target, compute, create, proxies, syslog, iso, util, debug, help} {
+	for _, f := range [][]cli.Flag{target, compute, create, cNetwork, memory, cpu, tls, registries, proxies, syslog, iso, util, debug, help} {
 		flags = append(flags, f...)
 	}
 
@@ -540,8 +516,16 @@ func (c *Create) processParams() error {
 		return err
 	}
 
-	if err := c.processContainerNetworks(); err != nil {
+	cNets, err := c.containerNetworks.ProcessContainerNetworks()
+	if err != nil {
 		return err
+	}
+	for i := range cNets {
+		alias := cNets[i].NetAlias
+		c.MappedNetworks[alias] = cNets[i].VNet
+		c.MappedNetworksGateways[alias] = cNets[i].Gateways
+		c.MappedNetworksIPRanges[alias] = cNets[i].IPRanges
+		c.MappedNetworksDNS[alias] = cNets[i].DNS
 	}
 
 	if err := c.processBridgeNetwork(); err != nil {
@@ -729,74 +713,6 @@ func (c *Create) processBridgeNetwork() error {
 	_, c.Data.BridgeIPRange, err = net.ParseCIDR(c.BridgeIPRange)
 	if err != nil {
 		return cli.NewExitError(fmt.Sprintf("Error parsing bridge network ip range: %s. Range must be in CIDR format, e.g., 172.16.0.0/12", err), 1)
-	}
-	return nil
-}
-
-func (c *Create) processContainerNetworks() error {
-	gws, err := parseContainerNetworkGateways([]string(c.containerNetworksGateway))
-	if err != nil {
-		return cli.NewExitError(err.Error(), 1)
-	}
-
-	pools, err := parseContainerNetworkIPRanges([]string(c.containerNetworksIPRanges))
-	if err != nil {
-		return cli.NewExitError(err.Error(), 1)
-	}
-
-	dns, err := parseContainerNetworkDNS([]string(c.containerNetworksDNS))
-	if err != nil {
-		return cli.NewExitError(err.Error(), 1)
-	}
-
-	// parse container networks
-	for _, cn := range c.containerNetworks {
-		vnet, v, err := splitVnetParam(cn)
-		if err != nil {
-			return cli.NewExitError(err.Error(), 1)
-		}
-
-		vicnet := vnet
-		if v != "" {
-			vicnet = v
-		}
-
-		c.MappedNetworks[vicnet] = vnet
-		c.MappedNetworksGateways[vicnet] = gws[vnet]
-		c.MappedNetworksIPRanges[vicnet] = pools[vnet]
-		c.MappedNetworksDNS[vicnet] = dns[vnet]
-
-		delete(gws, vnet)
-		delete(pools, vnet)
-		delete(dns, vnet)
-	}
-
-	var hasError bool
-	fmtMsg := "The following container network %s is set, but CONTAINER-NETWORK cannot be found. Please check the --container-network and %s settings"
-	if len(gws) > 0 {
-		log.Error(fmt.Sprintf(fmtMsg, "gateway", "--container-network-gateway"))
-		for key, value := range gws {
-			mask, _ := value.Mask.Size()
-			log.Errorf("\t%s:%s/%d, %q should be vSphere network name", key, value.IP, mask, key)
-		}
-		hasError = true
-	}
-	if len(pools) > 0 {
-		log.Error(fmt.Sprintf(fmtMsg, "ip range", "--container-network-ip-range"))
-		for key, value := range pools {
-			log.Errorf("\t%s:%s, %q should be vSphere network name", key, value, key)
-		}
-		hasError = true
-	}
-	if len(dns) > 0 {
-		log.Errorf(fmt.Sprintf(fmtMsg, "dns", "--container-network-dns"))
-		for key, value := range dns {
-			log.Errorf("\t%s:%s, %q should be vSphere network name", key, value, key)
-		}
-		hasError = true
-	}
-	if hasError {
-		return cli.NewExitError("Inconsistent container network configuration.", 1)
 	}
 	return nil
 }
@@ -1510,118 +1426,4 @@ func (c *Create) Run(clic *cli.Context) (err error) {
 	log.Infof("Installer completed successfully")
 
 	return nil
-}
-
-type ipNetUnmarshaler struct {
-	ipnet *net.IPNet
-	ip    net.IP
-}
-
-func (m *ipNetUnmarshaler) UnmarshalText(text []byte) error {
-	s := string(text)
-	ip, ipnet, err := net.ParseCIDR(s)
-	if err != nil {
-		return err
-	}
-
-	m.ipnet = ipnet
-	m.ip = ip
-	return nil
-}
-
-func parseContainerNetworkGateways(cgs []string) (map[string]net.IPNet, error) {
-	gws := make(map[string]net.IPNet)
-	for _, cg := range cgs {
-		m := &ipNetUnmarshaler{}
-		vnet, err := parseVnetParam(cg, m)
-		if err != nil {
-			return nil, err
-		}
-
-		if _, ok := gws[vnet]; ok {
-			return nil, fmt.Errorf("Duplicate gateway specified for container network %s", vnet)
-		}
-
-		gws[vnet] = net.IPNet{IP: m.ip, Mask: m.ipnet.Mask}
-	}
-
-	return gws, nil
-}
-
-func parseContainerNetworkIPRanges(cps []string) (map[string][]ip.Range, error) {
-	pools := make(map[string][]ip.Range)
-	for _, cp := range cps {
-		ipr := &ip.Range{}
-		vnet, err := parseVnetParam(cp, ipr)
-		if err != nil {
-			return nil, err
-		}
-
-		pools[vnet] = append(pools[vnet], *ipr)
-	}
-
-	return pools, nil
-}
-
-func parseContainerNetworkDNS(cds []string) (map[string][]net.IP, error) {
-	dns := make(map[string][]net.IP)
-	for _, cd := range cds {
-		var ip net.IP
-		vnet, err := parseVnetParam(cd, &ip)
-		if err != nil {
-			return nil, err
-		}
-
-		if ip == nil {
-			return nil, fmt.Errorf("DNS IP not specified for container network %s", vnet)
-		}
-
-		dns[vnet] = append(dns[vnet], ip)
-	}
-
-	return dns, nil
-}
-
-func splitVnetParam(p string) (vnet string, value string, err error) {
-	mapped := strings.Split(p, ":")
-	if len(mapped) == 0 || len(mapped) > 2 {
-		err = fmt.Errorf("Invalid value for parameter %s", p)
-		return
-	}
-
-	vnet = mapped[0]
-	if vnet == "" {
-		err = fmt.Errorf("Container network not specified in parameter %s", p)
-		return
-	}
-
-	// If the supplied vSphere network contains spaces then the user must supply a network alias. Guest info won't receive a name with spaces.
-	if strings.Contains(vnet, " ") && (len(mapped) == 1 || (len(mapped) == 2 && len(mapped[1]) == 0)) {
-		err = fmt.Errorf("A network alias must be supplied when network name %q contains spaces.", p)
-		return
-	}
-
-	if len(mapped) > 1 {
-		// Make sure the alias does not contain spaces
-		if strings.Contains(mapped[1], " ") {
-			err = fmt.Errorf("The network alias supplied in %q cannot contain spaces.", p)
-			return
-		}
-		value = mapped[1]
-	}
-
-	return
-}
-
-func parseVnetParam(p string, m encoding.TextUnmarshaler) (vnet string, err error) {
-	vnet, v, err := splitVnetParam(p)
-	if err != nil {
-		return "", fmt.Errorf("Error parsing container network parameter %s: %s", p, err)
-	}
-
-	if err = m.UnmarshalText([]byte(v)); err != nil {
-		return "", fmt.Errorf("Error parsing container network parameter %s: %s", p, err)
-	}
-
-	return vnet, nil
 }
