+++ tests/test-cases/Group6-VIC-Machine/6-16-Config.robot
@@ -73,4 +73,55 @@ Configure VCH https-proxy
     Should Not Contain  ${output}  proxy.vmware.com:3128
     ${output}=  Run  bin/vic-machine-linux inspect --name=%{VCH-NAME} --target="%{TEST_USERNAME}:%{TEST_PASSWORD}@%{TEST_URL}" --thumbprint=%{TEST_THUMBPRINT} --conf
     Should Contain  ${output}  --https-proxy=https://proxy.vmware.com:3128
-	Should Not Contain  ${output}  --http-proxy
+    Should Not Contain  ${output}  --http-proxy
+
+Configure VCH Container Networks
+    ${out}=  Run  govc host.portgroup.remove vm-network
+    ${out}=  Run  govc host.portgroup.add -vswitch vSwitchLAN vm-network
+
+    ${output}=  Run  bin/vic-machine-linux configure --name=%{VCH-NAME} --target=%{TEST_URL} --thumbprint=%{TEST_THUMBPRINT} --user=%{TEST_USERNAME} --password=%{TEST_PASSWORD} --timeout %{TEST_TIMEOUT} --container-network vm-network:vmnet
+    Should Contain  ${output}  Completed successfully
+
+    ${rc}  ${output}=  Run And Return Rc And Output  docker %{VCH-PARAMS} network ls
+    Should Be Equal As Integers  ${rc}  0
+    Should Contain  ${output}  vmnet
+
+    ${output}=  Run  bin/vic-machine-linux inspect --name=%{VCH-NAME} --target="%{TEST_USERNAME}:%{TEST_PASSWORD}@%{TEST_URL}" --thumbprint=%{TEST_THUMBPRINT} --conf
+    Should Contain  ${output}  --container-network=vm-network:vmnet
+
+    # Test that configure fails if an existing container-network is not specified
+    ${out}=  Run  govc host.portgroup.remove management
+    ${out}=  Run  govc host.portgroup.add -vswitch vSwitchLAN management
+    ${output}=  Run  bin/vic-machine-linux configure --name=%{VCH-NAME} --target=%{TEST_URL} --thumbprint=%{TEST_THUMBPRINT} --user=%{TEST_USERNAME} --password=%{TEST_PASSWORD} --timeout %{TEST_TIMEOUT} --container-network management:mgmt --container-network-ip-range=management:10.10.10.0/24 --container-network-gateway=management:10.10.10.1/24
+    Should Contain  ${output}  all existing container networks must also be specified
+    Should Not Contain  ${output}  Completed successfully
+
+    # Add another container network while specifying the existing one
+    ${output}=  Run  bin/vic-machine-linux configure --name=%{VCH-NAME} --target=%{TEST_URL} --thumbprint=%{TEST_THUMBPRINT} --user=%{TEST_USERNAME} --password=%{TEST_PASSWORD} --timeout %{TEST_TIMEOUT} --container-network vm-network:vmnet --container-network management:mgmt --container-network-ip-range=management:10.10.10.0/24 --container-network-gateway=management:10.10.10.1/24
+    Should Contain  ${output}  Completed successfully
+
+    ${rc}  ${output}=  Run And Return Rc And Output  docker %{VCH-PARAMS} network ls
+    Should Be Equal As Integers  ${rc}  0
+    Should Contain  ${output}  vmnet
+    Should Contain  ${output}  mgmt
+
+    ${output}=  Run  bin/vic-machine-linux inspect --name=%{VCH-NAME} --target="%{TEST_USERNAME}:%{TEST_PASSWORD}@%{TEST_URL}" --thumbprint=%{TEST_THUMBPRINT} --conf
+    Should Contain  ${output}  --container-network=vm-network:vmnet
+    Should Contain  ${output}  --container-network=management:mgmt
+    Should Contain  ${output}  --container-network-ip-range=management:10.10.10.0/24
+    Should Contain  ${output}  --container-network-gateway=management:10.10.10.1/24
+
+    # Test that changes to existing networks are not supported
+    ${output}=  Run  bin/vic-machine-linux configure --name=%{VCH-NAME} --target=%{TEST_URL} --thumbprint=%{TEST_THUMBPRINT} --user=%{TEST_USERNAME} --password=%{TEST_PASSWORD} --timeout %{TEST_TIMEOUT} --container-network vm-network:vmnet --container-network management:mgmt --container-network-ip-range=management:10.10.10.0/24 --container-network-gateway=management:10.10.10.2/24
+    Should Contain  ${output}  changes to existing container networks are not supported
+    Should Not Contain  ${output}  Completed successfully
+    ${output}=  Run  bin/vic-machine-linux configure --name=%{VCH-NAME} --target=%{TEST_URL} --thumbprint=%{TEST_THUMBPRINT} --user=%{TEST_USERNAME} --password=%{TEST_PASSWORD} --timeout %{TEST_TIMEOUT} --container-network vm-network:vmnet --container-network management:mgmt --container-network-ip-range=management:10.10.10.0/16 --container-network-gateway=management:10.10.10.1/24
+    Should Contain  ${output}  changes to existing container networks are not supported
+    Should Not Contain  ${output}  Completed successfully
+    ${output}=  Run  bin/vic-machine-linux configure --name=%{VCH-NAME} --target=%{TEST_URL} --thumbprint=%{TEST_THUMBPRINT} --user=%{TEST_USERNAME} --password=%{TEST_PASSWORD} --timeout %{TEST_TIMEOUT} --container-network vm-network:vmnet --container-network management:mgmt
+    Should Contain  ${output}  changes to existing container networks are not supported
+    Should Not Contain  ${output}  Completed successfully
+
+    # Clean up portgroups
+    ${out}=  Run  govc host.portgroup.remove vm-network
+    ${out}=  Run  govc host.portgroup.remove management
